Finishing mapping

+++ src/Hann.Assesment/Core/Hahn.Assesment.Domain/Entities/SeverityAlert.cs
@@ -4,8 +4,8 @@ public class SeverityAlert
 {
     public Guid Id { get; set; }
     public DateTime UpdatedAt { get; set; }
-    public long Start { get; set; }
-    public long End { get; set; }
+    public int Start { get; set; }
+    public int End { get; set; }
     public int WindowsSizeHours { get; set; }
 
     public virtual IEnumerable<SeverityReport>? SeverityReports { get; set; }
